Clean up warnings part 1 of ? (#434)
* Fix various compilation warnings.

* Clean up warning in event classes.  Removed unused parameter from navigation event constructor.

* Fixed warnings in annotation classes

* Fix type related warning

* Fix warnings in settings classes.

* Use draw list that is passed in to sticky and render to it.
Make button invisible otherwise it renders over the draw list.
Shadows now appear since draw list isn't clipped by overlayed window anymore.

Clamp sticky (annotation)  position when dragging so that it can't be dragged out of window bounds.  When that happens it can never be retrieved.

* Fix for: warning: ISO C++ forbids converting a string constant to 'char*'

* Add include for clamp (<algorithm>)
Add missing copyright.

* fix non const pointer

* Show formatted time on tooltip.
Fix variable name LineColor -> line_color.
Added line_thickness variable.

+++ src/view/src/rocprofvis_event_search.cpp
@@ -52,7 +52,7 @@ EventSearch::Update()
                i != m_time_column_indices[kTimeStartNs] &&
                i != m_time_column_indices[kDurationNs])
             {
-                m_hidden_column_indices.push_back(i);
+                m_hidden_column_indices.push_back(static_cast<int>(i));
             }
         }
     }
